move uploader and downloader to file packege. Fix circular reference
